Safari-609.1.11.1/LayoutTests/http/tests/security/cross-frame-access-custom-expected.txt

CONSOLE MESSAGE: line 57: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 86: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 86: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 86: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 86: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 86: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 86: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 86: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 114: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.


----- tests for getting/setting interesting properties -----

PASS: canGet('targetWindow.Object') should be 'false' and is.
PASS: canSet('targetWindow.Object') should be 'false' and is.
PASS: canGet('targetWindow.existingCustomProperty') should be 'false' and is.
PASS: canSet('targetWindow.existingCustomProperty') should be 'false' and is.
PASS: canSet('targetWindow.newCustomProperty') should be 'false' and is.
PASS: canGet('targetWindow.hasOwnProperty') should be 'false' and is.
PASS: canSet('targetWindow.hasOwnProperty') should be 'false' and is.
PASS: canGet('targetWindow.windowPrototypeCustomProperty') should be 'false' and is.
PASS: canSet('targetWindow.windowPrototypeCustomProperty') should be 'false' and is.
PASS: canGet('targetWindow.objectPrototypeCustomProperty') should be 'false' and is.
PASS: canSet('targetWindow.objectPrototypeCustomProperty') should be 'false' and is.
PASS: canGet('targetWindow.toString') should be 'false' and is.
PASS: canSet('targetWindow.toString') should be 'false' and is.
PASS: toString('targetWindow', 'NOTALLOWED') should be 'NOTALLOWED' and is.
